% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_huishoudens.R
\name{bepaal_huishoudens}
\alias{bepaal_huishoudens}
\title{Hoofd functie om 'huishouden' kolom toe te voegen.}
\usage{
bepaal_huishoudens(
  peil_datum,
  brpstam,
  historie,
  huwelijk,
  kind,
  adressen_inst,
  verhuis_wezen = TRUE,
  ethniciteit = TRUE,
  buurt_wijk_codes = TRUE,
  ...
)
}
\description{
Hoofd functie om 'huishouden' kolom toe te voegen.
}
