% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_huishoudens.R
\name{brp_summary}
\alias{brp_summary}
\title{Huishouden data samenvatting}
\usage{
brp_summary(data)
}
\description{
Huishouden data samenvatting
}
\details{
Werkt met output van bepaal_huishoudens
}
