% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_huishoudens.R
\name{brp_tijdmachine}
\alias{brp_tijdmachine}
\title{Bepaal inwonerbestand op een willekeurige datum}
\usage{
brp_tijdmachine(historie, brpstam, peil_datum)
}
\arguments{
\item{historie}{Dataframe ingelezen met}
}
\description{
Bepaal inwonerbestand op een willekeurige datum
}
