% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_huishoudens.R
\name{brp_tijdmachine_cipers}
\alias{brp_tijdmachine_cipers}
\title{BRP tijdmachine, ouder CIPERS levering}
\usage{
brp_tijdmachine_cipers(historie, brpstam, peil_datum)
}
\arguments{
\item{historie}{Dataframe ingelezen met read_historie}
}
\description{
BRP tijdmachine, ouder CIPERS levering
}
