% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_brpstam}
\alias{read_brpstam}
\title{Lees stambestand (BZSPRSQ00)}
\usage{
read_brpstam(
  brp_bzsprs,
  adressen_inst,
  peil_datum,
  date_format = c("new", "old")
)
}
\description{
Lees stambestand (BZSPRSQ00)
}
