% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_institutionele_adressen}
\alias{read_institutionele_adressen}
\title{Lees institutionele adressen data}
\usage{
read_institutionele_adressen(
  con = NULL,
  inst_path = NULL,
  basename = "adressen_lijst.csv"
)
}
\description{
Lees institutionele adressen data
}
